# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/surveys": {
      "get": {
        "operationId": "Surveys_list",
        "parameters": [
          {
            "in": "query",
            "name": "box",
            "required": false,
            "schema": {
              "enum": [
                "mine",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Survey: le mie (invitato) oppure tutte (HR, box=all)",
        "tags": [
          "surveys"
        ]
      },
      "post": {
        "operationId": "Surveys_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "anonymityThreshold": {
                    "default": 5,
                    "maximum": 50,
                    "minimum": 3,
                    "type": "integer"
                  },
                  "anonymous": {
                    "default": true,
                    "type": "boolean"
                  },
                  "closesAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "description": {
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "drivers": {
                    "additionalProperties": {
                      "type": "string"
                    },
                    "type": "object"
                  },
                  "enpsField": {
                    "maxLength": 60,
                    "nullable": true,
                    "type": "string"
                  },
                  "formKey": {
                    "pattern": "^[a-z][a-z0-9_]{0,60}$",
                    "type": "string"
                  },
                  "population": {
                    "additionalProperties": false,
                    "default": {},
                    "properties": {
                      "excludePersonIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "orgUnitIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "personIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "rotation": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "template": {
                    "enum": [
                      "engagement",
                      "pulse",
                      "enps",
                      "wellbeing"
                    ],
                    "type": "string"
                  },
                  "title": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea una survey da template della libreria (engagement, pulse, eNPS, benessere) o da un form pubblicato",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}": {
      "get": {
        "operationId": "Surveys_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "surveys"
        ]
      },
      "patch": {
        "operationId": "Surveys_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "anonymityThreshold": {
                    "maximum": 50,
                    "minimum": 3,
                    "type": "integer"
                  },
                  "anonymous": {
                    "type": "boolean"
                  },
                  "closesAt": {
                    "format": "date-time",
                    "nullable": true,
                    "type": "string"
                  },
                  "description": {
                    "maxLength": 2000,
                    "nullable": true,
                    "type": "string"
                  },
                  "population": {
                    "additionalProperties": false,
                    "properties": {
                      "excludePersonIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "orgUnitIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      },
                      "personIds": {
                        "items": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "type": "array"
                      }
                    },
                    "type": "object"
                  },
                  "title": {
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/close": {
      "post": {
        "operationId": "Surveys_close",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/extend": {
      "post": {
        "operationId": "Surveys_extend",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "closesAt": {
                    "format": "date-time",
                    "type": "string"
                  }
                },
                "required": [
                  "closesAt"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/form": {
      "get": {
        "operationId": "Surveys_form",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Questionario da compilare (solo invitati) con dichiarazione di anonimato",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/launch": {
      "post": {
        "operationId": "Surveys_launch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "closesAt": {
                    "format": "date-time",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lancia: crea gli inviti e notifica la popolazione",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/remind": {
      "post": {
        "operationId": "Surveys_remind",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sollecita i non rispondenti (restituisce solo i conteggi)",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/respond": {
      "post": {
        "operationId": "Surveys_respond",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "answers": {
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        },
                        {
                          "type": "boolean"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        {
                          "enum": [
                            "null"
                          ],
                          "nullable": true
                        }
                      ]
                    },
                    "type": "object"
                  }
                },
                "required": [
                  "answers"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invia le risposte: nelle survey anonime nessun legame con la persona",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/results": {
      "get": {
        "operationId": "Surveys_results",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "segment",
            "required": false,
            "schema": {
              "default": "org_unit",
              "enum": [
                "org_unit",
                "manager",
                "tenure"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Risultati aggregati con soglia di anonimato: driver, domande, eNPS, heatmap, commenti (HR), confronto con la precedente",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/share": {
      "post": {
        "operationId": "Surveys_share",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "summary": {
                    "maxLength": 5000,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "summary"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pubblica la sintesi ai rispondenti",
        "tags": [
          "surveys"
        ]
      }
    },
    "/api/v1/surveys/{id}/summary": {
      "get": {
        "operationId": "Surveys_summary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Errore (RFC 9457)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sintesi pubblicata dall’HR (dopo la condivisione)",
        "tags": [
          "surveys"
        ]
      }
    },
...
  },
  "info": {
    "title": "WorkingBetter API",
    "description": "API REST unica per web, mobile e connettori (ADR-0005). Errori in formato RFC 9457 (application/problem+json). I body e le query sono descritti dagli stessi schemi Zod usati per la validazione (ADR-0009).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "description": "URI del tipo di errore (about:blank per gli errori generici)"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "description": "Identificativo della richiesta (x-request-id)"
          },
          "code": {
            "type": "string",
            "description": "Codice applicativo stabile (ErrorCodes in @wb/shared)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "path",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}